task2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.0.0
evaluate>=0.4.0
seqeval>=1.2.2
numpy>=1.24.0
Pillow>=9.0.0
kagglehub>=0.1.0
